move to spark package; get refresh interval
